fix(web): treat assignment runtime as a closed contract object
Code review found that preserving unknown `runtime` sub-keys on edit writes an
assignments.json the CLI rejects (RuntimeRef is decoded strictly with no Extra;
schema additionalProperties:false), wedging the whole file. Drop that
preservation so an edit rebuilds runtime from the known sub-keys and a foreign
key self-heals away. Also: keep runs-on editable in container mode (a container
can target a specific runner) and hard-block only the container + macOS/Windows
combination the runner rejects; delete the now-orphaned containerRunnerWarning;
split the field components into AdvancedRuntimeFields.tsx to get the form back
under 1k lines; and add runtime.ts unit tests plus reject-branch coverage.
