+++ web/src/api/mutations/assignments.ts
@@ -21,6 +21,7 @@ import { parseRunnerLabels } from "@/util/runners"
 import {
   RUNTIME_LANGUAGES,
   type RuntimeLanguage,
+  isNonUbuntuHostedLabel,
   parseAptPackages,
   validateAptPackages,
   validateLanguageVersion,
@@ -765,6 +766,15 @@ async function buildAssignmentEntry(
     runtime["runs-on"] = runnerLabels
   }
   if (containerImage) {
+    // Containers run on Ubuntu hosts only — reject a macOS/Windows runs-on
+    // label, matching the CLI's ValidateRuntime (a custom/self-hosted or Ubuntu
+    // label is fine, so a container can still target a specific runner).
+    const badLabel = runnerLabels.find(isNonUbuntuHostedLabel)
+    if (badLabel) {
+      throw new Error(
+        `runtime.runs-on ${JSON.stringify(badLabel)} can't be combined with a Docker image — GitHub Actions runs containers on Ubuntu hosts only.`,
+      )
+    }
     runtime.container = { image: containerImage }
     if (containerUser) {
       runtime.container.user = containerUser
@@ -1297,6 +1307,10 @@ const ASSIGNMENT_KEY_OWNERSHIP: Record<
   autograder: "managed",
   max_group_size: "managed",
   feedback_pr: "managed",
+  // Fully managed AND a closed object: the CLI decodes runtime strictly
+  // (RuntimeRef has no Extra, DisallowUnknownFields; schema additionalProperties
+  // false), so the rebuilt runtime must win and any unknown sub-key drops rather
+  // than round-tripping into a file the CLI would reject.
   runtime: "managed",
   allowed_files: "managed",
   pass_threshold: "managed",
@@ -1314,44 +1328,16 @@ const EDIT_MANAGED_ASSIGNMENT_KEYS = new Set<string>(
     .map(([key]) => key),
 )
 
-// Runtime sub-keys the edit form fully manages (rebuilt from input, so a
-// clearing edit wins). Anything else inside `runtime` (an unknown future
-// sub-key) is preserved verbatim on edit — the same "tolerate AND preserve"
-// contract applied at the assignment level, but one level deeper, since the
-// form rebuilds the whole `runtime` object from scratch. Keeping this in lock-
-// step with the CLI's RuntimeRef guards teachers who mix CLI/manual runtime
-// config with GUI edits.
-const RUNTIME_MANAGED_KEYS = new Set<string>([
-  "runs-on",
-  "container",
-  "python",
-  "node",
-  "java",
-  "go",
-  "apt",
-])
-
-// Copy forward unknown `runtime` sub-keys from the existing entry onto the
-// rebuilt runtime, without touching managed sub-keys. Returns the (possibly
-// new) runtime object, or undefined when neither side has one. Pure.
-export function preserveUnmanagedRuntimeKeys(
-  existingRuntime: Record<string, unknown> | undefined,
-  editedRuntime: Record<string, unknown> | undefined,
-): Record<string, unknown> | undefined {
-  if (!existingRuntime) return editedRuntime
-  const carried: Record<string, unknown> = {}
-  for (const [key, value] of Object.entries(existingRuntime)) {
-    if (RUNTIME_MANAGED_KEYS.has(key)) continue
-    if (value === undefined) continue
-    carried[key] = value
-  }
-  if (Object.keys(carried).length === 0) return editedRuntime
-  return { ...(editedRuntime ?? {}), ...carried }
-}
-
 // Copy forward entry-level keys the edit form doesn't manage (e.g.
 // `migrated_from`, unknown future keys) onto the rebuilt edit, without
 // overwriting managed keys. Mirrors the CLI's AssignmentEntry.Extra round-trip.
+//
+// `runtime` is deliberately NOT preserved this way: it's a managed key, and the
+// CLI decodes it as a CLOSED object (RuntimeRef has no Extra, decoded with
+// DisallowUnknownFields; the schema sets additionalProperties:false). Carrying
+// an unknown runtime sub-key forward would write an assignments.json the CLI
+// refuses to parse. So an edit rebuilds runtime from the known sub-keys and any
+// foreign key self-heals away — matching the CLI's own strictness.
 export function preserveUnmanagedAssignmentKeys(
   existing: Assignment,
   edited: Assignment,
@@ -1364,21 +1350,6 @@ export function preserveUnmanagedAssignmentKeys(
     if (value === undefined) continue
     merged[key] = value
   }
-
-  // `runtime` is a managed key (a clearing edit must win), but the form only
-  // knows the schema's runtime sub-keys, so carry forward any unknown ones from
-  // the existing entry so a GUI edit doesn't silently drop CLI/manual runtime
-  // config it can't model yet.
-  const nextRuntime = preserveUnmanagedRuntimeKeys(
-    existing.runtime as Record<string, unknown> | undefined,
-    edited.runtime as Record<string, unknown> | undefined,
-  )
-  if (nextRuntime && Object.keys(nextRuntime).length > 0) {
-    merged.runtime = nextRuntime
-  } else {
-    delete merged.runtime
-  }
-
   return merged as Assignment
 }
 // grant — the same write + grant as createAssignment, minus form resolution.
