+++ web/src/pages/assignments/CreateAssignmentForm.tsx
@@ -1,18 +1,9 @@
 import { useForm } from "@tanstack/react-form"
-import { useQuery } from "@tanstack/react-query"
 import { useState } from "react"
 import { useTranslation } from "react-i18next"
 import type { TFunction } from "i18next"
 import { slugify } from "@/util/slug"
-import {
-  AlertTriangle,
-  Check,
-  CheckCircle2,
-  ChevronDown,
-  HelpCircle,
-  Loader2,
-  ServerCog,
-} from "lucide-react"
+import { AlertTriangle, HelpCircle } from "lucide-react"
 import AutogradingTestsPane from "./AutogradingTestsPane"
 import type { AssignmentTestDraft } from "@/util/assignmentTests"
 import {
@@ -25,31 +16,29 @@ import {
   allowedFilesToText,
   validateAllowedFiles,
 } from "@/util/allowedFiles"
-import {
-  isRunnerLabelShapeValid,
-  verifyRunnerLabels,
-  parseRunnerLabels,
-  isKnownHostedRunnerLabel,
-  isStandardSelfHostedLabel,
-  type OrgRunnersResult,
-  type RunnerVerification,
-} from "@/util/runners"
-import { orgRunnersQuery } from "@/hooks/github/queries"
+import { parseRunnerLabels } from "@/util/runners"
 import {
   RUNTIME_LANGUAGES,
-  RUNTIME_LANGUAGE_META,
   type RuntimeLanguage,
   aptPackagesToText,
+  isNonUbuntuHostedLabel,
   parseAptPackages,
   validateAptPackages,
   validateLanguageVersion,
 } from "@/util/runtime"
-import { useOptionalGitHubClient } from "@/context/github/GitHubProvider"
 import { TemplateField } from "./TemplateField"
 import {
-  useDebouncedValue,
+  FieldLabel,
+  RunnerField,
+  LanguageVersionField,
+  ContainerFields,
+  AptField,
+} from "./AdvancedRuntimeFields"
+import {
   normalizeOnBlur,
-  type StringField,
+  toDatetimeLocalValue,
+  sevenDaysFromNow,
+  utcIsoToDatetimeLocalValue,
 } from "./formFieldHelpers"
 import type { Assignment } from "@/types/classroom"
 import { GROUP_SIZE_MAX, GROUP_SIZE_MIN } from "@/types/classroom"
@@ -235,15 +224,34 @@ const useAssignmentForm = (
           }
         }
 
+        // A container runs on Ubuntu hosts only, so a macOS/Windows runner
+        // label can't be combined with a Docker image (mirrors the CLI). A
+        // custom/self-hosted or Ubuntu label is fine.
+        if (value.runtime_env === "container" && value.container_image.trim()) {
+          const badLabel = parseRunnerLabels(value.runs_on).find(
+            isNonUbuntuHostedLabel,
+          )
+          if (badLabel) {
+            errors.runs_on = t(
+              "assignments.form.runtime.runnerContainerError",
+              {
+                label: badLabel,
+              },
+            )
+          }
+        }
+
         return Object.keys(errors).length > 0 ? { fields: errors } : undefined
       },
     },
     onSubmit: async ({ value }) => {
       // Clear the fields that don't belong to the selected runtime environment
-      // so a hidden, stale value from the other mode can't reach the wire —
-      // this is what makes the container-vs-(runs-on/apt) conflict
-      // unrepresentable, not just validated-against. Language versions apply to
-      // both modes (setup-* runs inside a container too), so they pass through.
+      // so a hidden, stale value from the other mode can't reach the wire.
+      // apt is hosted-only (a container image owns its packages — the CLI
+      // forbids container+apt), and container image/user apply only in
+      // container mode. runs-on and the language versions apply to BOTH modes
+      // (a container job can target a specific runner; setup-* runs inside a
+      // container), so they always pass through.
       const isContainer = value.runtime_env === "container"
       await onSubmit({
         name: value.name.trim(),
@@ -255,7 +263,7 @@ const useAssignmentForm = (
         max_group_size: value.max_group_size,
         feedback_pr: value.feedback_pr,
         runtime_env: value.runtime_env,
-        runs_on: isContainer ? "" : value.runs_on.trim(),
+        runs_on: value.runs_on.trim(),
         container_image: isContainer ? value.container_image.trim() : "",
         container_user: isContainer ? value.container_user.trim() : "",
         runtime_python: value.runtime_python.trim(),
@@ -303,479 +311,6 @@ const FormErrors = ({ form }: { form: AssignmentForm }) => (
     )}
   </form.Subscribe>
 )
-
-// A bold field label with an optional help affordance: a question-mark icon
-// that reveals detailed guidance on hover/focus (DaisyUI tooltip, theme-aware).
-// Keeps the label short so the form stays scannable while the "why/how" moves
-// into the tooltip. `help` is the tooltip text; `htmlFor` ties the label to its
-// control. The icon is a real focusable button (keyboard/screen-reader
-// reachable) carrying the same text as its accessible name.
-const FieldLabel = ({
-  htmlFor,
-  label,
-  help,
-  required,
-}: {
-  htmlFor?: string
-  label: string
-  help?: string
-  required?: boolean
-}) => (
-  <div className="mb-1.5 flex items-center gap-1.5">
-    <label htmlFor={htmlFor} className="label font-bold">
-      {label}
-      {required ? <span className="text-error">*</span> : null}
-    </label>
-    {help ? (
-      <span
-        className="tooltip tooltip-bottom before:max-w-xs before:whitespace-normal before:text-left"
-        data-tip={help}
-      >
-        <button
-          type="button"
-          aria-label={help}
-          className="btn btn-ghost btn-xs btn-circle text-base-content/50 hover:text-base-content"
-        >
-          <HelpCircle aria-hidden="true" className="size-4" />
-        </button>
-      </span>
-    ) : null}
-  </div>
-)
-
-// A language toolchain version input (python/node/java/go). A themed combobox:
-// a text input with a chevron that opens a DaisyUI dropdown of the actively-
-// supported versions, but the input stays free-text so a teacher can type any
-// custom version. Empty = toolchain off (except Python, which the runner
-// defaults to 3.12). Advisory shape check mirrors the CLI's
-// LanguageVersionPattern.
-const LanguageVersionField = ({
-  form,
-  language,
-}: {
-  form: AssignmentForm
-  language: RuntimeLanguage
-}) => {
-  const { t } = useTranslation()
-  const fieldName = `runtime_${language}` as const
-  const meta = RUNTIME_LANGUAGE_META[language]
-  return (
-    <form.Field name={fieldName}>
-      {(field) => {
-        const error = field.state.meta.errors[0] as string | undefined
-        const current = field.state.value.trim()
-        // Close the focus-driven dropdown after a pick so it doesn't linger
-        // (DaisyUI dropdowns stay open while a descendant holds focus).
-        const closeDropdown = () =>
-          (document.activeElement as HTMLElement | null)?.blur()
-        return (
-          <div>
-            <FieldLabel
-              htmlFor={field.name}
-              label={meta.label}
-              help={t("assignments.form.runtime.versionTip", {
-                language: meta.label,
-              })}
-            />
-            <div className="dropdown w-full max-w-xs">
-              <div className="join w-full">
-                <input
-                  id={field.name}
-                  name={field.name}
-                  type="text"
-                  autoComplete="off"
-                  spellCheck={false}
-                  className="input join-item w-full"
-                  placeholder={t(
-                    "assignments.form.runtime.versionPlaceholder",
-                    { version: meta.placeholder },
-                  )}
-                  value={field.state.value}
-                  onBlur={normalizeOnBlur(field)}
-                  onChange={(e) => field.handleChange(e.target.value)}
-                />
-                <button
-                  type="button"
-                  tabIndex={0}
-                  className="btn btn-square join-item border-base-content/20"
-                  aria-label={t("assignments.form.runtime.versionMenu", {
-                    language: meta.label,
-                  })}
-                >
-                  <ChevronDown aria-hidden="true" className="size-4" />
-                </button>
-              </div>
-              <ul
-                tabIndex={0}
-                className="dropdown-content menu z-10 mt-1 w-full rounded-box border border-base-content/5 bg-base-100 p-1 shadow"
-              >
-                {meta.versions.map((version) => (
-                  <li key={version}>
-                    <button
-                      type="button"
-                      className={
-                        version === current ? "active font-semibold" : undefined
-                      }
-                      onClick={() => {
-                        field.handleChange(version)
-                        closeDropdown()
-                      }}
-                    >
-                      <Check
-                        aria-hidden="true"
-                        className={`size-4 ${
-                          version === current ? "" : "invisible"
-                        }`}
-                      />
-                      {version}
-                    </button>
-                  </li>
-                ))}
-              </ul>
-            </div>
-            {error ? (
-              <p
-                role="alert"
-                className="mt-1.5 flex items-center gap-1.5 text-sm text-error"
-              >
-                <AlertTriangle aria-hidden="true" className="size-4 shrink-0" />
-                {error}
-              </p>
-            ) : null}
-          </div>
-        )
-      }}
-    </form.Field>
-  )
-}
-
-// Free-form runner input with advisory, non-blocking verification: annotates
-// the value but never rewrites or clears what the teacher typed.
-const RunnerField = ({ field, org }: { field: StringField; org?: string }) => {
-  const { t } = useTranslation()
-  const client = useOptionalGitHubClient()
-  const rawValue = field.state.value
-  const debouncedValue = useDebouncedValue(rawValue.trim(), 400)
-
-  // Hit the org runners API only for a well-shaped label not already recognized
-  // client-side; everything else needs no network call.
-  const needsOrgLookup = Boolean(
-    client &&
-    org &&
-    parseRunnerLabels(debouncedValue).some(
-      (label) =>
-        isRunnerLabelShapeValid(label) &&
-        !isKnownHostedRunnerLabel(label) &&
-        !isStandardSelfHostedLabel(label),
-    ),
-  )
-
-  const orgRunnersResultQuery = useQuery({
-    ...orgRunnersQuery(client!, org ?? ""),
-    enabled: needsOrgLookup,
-  })
-
-  const orgRunners: OrgRunnersResult = needsOrgLookup
-    ? (orgRunnersResultQuery.data ?? { available: false, reason: "error" })
-    : { available: false, reason: "no-access" }
-
-  // Hold off on the "not found" verdict while the lookup is in flight.
-  const isVerifying = needsOrgLookup && orgRunnersResultQuery.isLoading
-
-  const pending = rawValue.trim() !== debouncedValue
-  const verification = verifyRunnerLabels(debouncedValue, orgRunners)
-
-  return (
-    <div>
-      <FieldLabel
-        htmlFor={field.name}
-        label={t("assignments.form.runner.label")}
-        help={t("assignments.form.runner.tip")}
-      />
-      <input
-        id={field.name}
-        name={field.name}
-        type="text"
-        autoComplete="off"
-        spellCheck={false}
-        className="input w-full max-w-xs"
-        placeholder="ubuntu-latest"
-        value={rawValue}
-        onBlur={normalizeOnBlur(field, (value) =>
-          parseRunnerLabels(value).join(", "),
-        )}
-        onChange={(e) => field.handleChange(e.target.value)}
-      />
-
-      <RunnerVerificationNote
-        verification={verification}
-        pending={pending || isVerifying}
-        hasValue={verification.kind !== "empty"}
-      />
-
-      {verification.kind === "self-hosted" && (
-        <p className="mt-1.5 text-xs text-base-content/70">
-          {t("assignments.form.runner.selfHostedHint_prefix")}{" "}
-          <code>self-hosted, linux, x64</code>
-          {t("assignments.form.runner.selfHostedHint_suffix")}
-        </p>
-      )}
-    </div>
-  )
-}
-
-// Container-runtime fields (Docker image + optional user). Rendered only in
-// container mode, so a container image can never coexist with runs-on/apt in
-// the UI — the conflict the runner rejects is unrepresentable here.
-const ContainerFields = ({ form }: { form: AssignmentForm }) => {
-  const { t } = useTranslation()
-  return (
-    <div className="grid grid-cols-1 gap-x-8 gap-y-4 sm:grid-cols-2">
-      <form.Field name="container_image">
-        {(field) => (
-          <div>
-            <FieldLabel
-              htmlFor={field.name}
-              label={t("assignments.form.dockerImage")}
-              help={t("assignments.form.dockerImageTip")}
-            />
-            <input
-              id={field.name}
-              name={field.name}
-              type="text"
-              className="input w-full max-w-xs"
-              placeholder={t("assignments.form.dockerImagePlaceholder")}
-              value={field.state.value}
-              onBlur={normalizeOnBlur(field)}
-              onChange={(e) => field.handleChange(e.target.value)}
-            />
-          </div>
-        )}
-      </form.Field>
-
-      <form.Field name="container_user">
-        {(field) => (
-          <div>
-            <FieldLabel
-              htmlFor={field.name}
-              label={t("assignments.form.containerUser")}
-              help={t("assignments.form.containerUserTip")}
-            />
-            <input
-              id={field.name}
-              name={field.name}
-              type="text"
-              className="input w-full max-w-xs"
-              placeholder={t("assignments.form.containerUserPlaceholder")}
-              value={field.state.value}
-              onBlur={normalizeOnBlur(field)}
-              onChange={(e) => field.handleChange(e.target.value)}
-            />
-          </div>
-        )}
-      </form.Field>
-    </div>
-  )
-}
-
-// Extra apt packages input. Rendered only in hosted mode (a container image
-// owns its own packages), so apt-with-container can't be expressed.
-const AptField = ({ form }: { form: AssignmentForm }) => {
-  const { t } = useTranslation()
-  return (
-    <form.Field name="runtime_apt">
-      {(field) => {
-        const packages = parseAptPackages(field.state.value)
-        const error = field.state.meta.errors[0] as string | undefined
-        return (
-          <div className="mt-4">
-            <FieldLabel
-              htmlFor={field.name}
-              label={t("assignments.form.runtime.aptLabel")}
-              help={t("assignments.form.runtime.aptTip")}
-            />
-            <input
-              id={field.name}
-              name={field.name}
-              type="text"
-              autoComplete="off"
-              spellCheck={false}
-              className="input w-full"
-              placeholder={t("assignments.form.runtime.aptPlaceholder")}
-              value={field.state.value}
-              onBlur={normalizeOnBlur(field, (value) =>
-                parseAptPackages(value).join(", "),
-              )}
-              onChange={(e) => field.handleChange(e.target.value)}
-            />
-            {error ? (
-              <p
-                role="alert"
-                className="mt-1.5 flex items-center gap-1.5 text-sm text-error"
-              >
-                <AlertTriangle aria-hidden="true" className="size-4 shrink-0" />
-                {error}
-              </p>
-            ) : (
-              packages.length > 0 && (
-                <p className="mt-1.5 text-xs text-base-content/70">
-                  {t("assignments.form.runtime.aptCount", {
-                    count: packages.length,
-                  })}
-                </p>
-              )
-            )}
-          </div>
-        )
-      }}
-    </form.Field>
-  )
-}
-
-const RunnerVerificationNote = ({
-  verification,
-  pending,
-  hasValue,
-}: {
-  verification: RunnerVerification
-  pending: boolean
-  hasValue: boolean
-}) => {
-  const { t } = useTranslation()
-  if (pending && hasValue) {
-    return (
-      <p className="mt-1.5 flex items-center gap-1.5 text-sm text-base-content/70">
-        <Loader2 aria-hidden="true" className="size-4 shrink-0 animate-spin" />
-        {t("assignments.form.runner.checking")}
-      </p>
-    )
-  }
-
-  switch (verification.kind) {
-    case "empty":
-      return (
-        <p className="mt-1.5 text-sm text-base-content/70">
-          {t("assignments.form.runner.emptyHint_prefix")}{" "}
-          <code>ubuntu-latest</code>
-          {t("assignments.form.runner.emptyHint_suffix")}
-        </p>
-      )
-
-    case "hosted":
-      return (
-        <p className="mt-1.5 flex items-center gap-1.5 text-sm text-success">
-          <CheckCircle2 aria-hidden="true" className="size-4 shrink-0" />
-          {t("assignments.form.runner.hosted")}
-        </p>
-      )
-
-    case "self-hosted": {
-      const matched = verification.labels.filter(
-        (l) => l.kind === "self-hosted-match",
-      )
-      const matchNames = matched.flatMap((l) =>
-        l.kind === "self-hosted-match" ? l.runnerNames : [],
-      )
-      const uniqueNames = Array.from(new Set(matchNames))
-      return (
-        <p className="mt-1.5 flex items-center gap-1.5 text-sm text-success">
-          <ServerCog aria-hidden="true" className="size-4 shrink-0" />
-          {verification.confirmed && uniqueNames.length > 0
-            ? t("assignments.form.runner.selfHostedMatch", {
-                count: uniqueNames.length,
-                names: `${uniqueNames.slice(0, 3).join(", ")}${
-                  uniqueNames.length > 3 ? "…" : ""
-                }`,
-              })
-            : t("assignments.form.runner.selfHostedLabels")}
-        </p>
-      )
-    }
-
-    case "problem": {
-      const badShape = verification.labels
-        .filter((l) => l.kind === "invalid-shape")
-        .map((l) => l.label)
-      const unverified = verification.labels
-        .filter((l) => l.kind === "unverified")
-        .map((l) => l.label)
-      const parts: string[] = []
-      if (badShape.length > 0) {
-        parts.push(
-          t("assignments.form.runner.invalidLabel", {
-            labels: badShape.map((l) => `"${l}"`).join(", "),
-          }),
-        )
-      }
-      if (unverified.length > 0) {
-        parts.push(
-          t("assignments.form.runner.noRunnerMatch", {
-            labels: unverified.map((l) => `"${l}"`).join(", "),
-          }),
-        )
-      }
-      return (
-        <p className="mt-1.5 flex items-center gap-1.5 text-sm text-error">
-          <AlertTriangle aria-hidden="true" className="size-4 shrink-0" />
-          {parts.join(" ")}
-        </p>
-      )
-    }
-
-    case "too-many":
-      return (
-        <p className="mt-1.5 flex items-center gap-1.5 text-sm text-error">
-          <AlertTriangle aria-hidden="true" className="size-4 shrink-0" />
-          {t("assignments.form.runner.tooMany", {
-            count: verification.count,
-          })}
-        </p>
-      )
-
-    case "unknown":
-      return (
-        <p className="mt-1.5 flex items-center gap-1.5 text-sm text-base-content/70">
-          <HelpCircle aria-hidden="true" className="size-4 shrink-0" />
-          {t("assignments.form.runner.cannotVerify")}
-        </p>
-      )
-
-    default:
-      return null
-  }
-}
-
-const toDatetimeLocalValue = (date: Date) => {
-  const pad = (value: number) => String(value).padStart(2, "0")
-
-  const year = date.getFullYear()
-  const month = pad(date.getMonth() + 1)
-  const day = pad(date.getDate())
-  const hours = pad(date.getHours())
-  const minutes = pad(date.getMinutes())
-
-  return `${year}-${month}-${day}T${hours}:${minutes}`
-}
-
-// Create-mode default: a week out gives students a sensible runway and avoids
-// the form defaulting to an already-overdue "now".
-const sevenDaysFromNow = () => {
-  const date = new Date()
-  date.setDate(date.getDate() + 7)
-  return date
-}
-
-const utcIsoToDatetimeLocalValue = (value?: string) => {
-  if (!value) return ""
-
-  const date = new Date(value)
-
-  if (Number.isNaN(date.getTime())) {
-    return ""
-  }
-
-  return toDatetimeLocalValue(date)
-}
 
 // Map a stored classroom50/assignments/v1 entry back into form values:
 // template as `owner/repo`, due as datetime-local, runtime split into
@@ -1203,17 +738,17 @@ const CreateAssignmentForm = ({
                 )}
               </form.Field>
 
+              <div className="grid grid-cols-1 gap-x-8 gap-y-4 sm:grid-cols-2">
+                <form.Field name="runs_on">
+                  {(field) => <RunnerField field={field} org={org} />}
+                </form.Field>
+              </div>
+
               <form.Subscribe selector={(state) => state.values.runtime_env}>
                 {(runtimeEnv) =>
                   runtimeEnv === "container" ? (
                     <ContainerFields form={form} />
-                  ) : (
-                    <div className="grid grid-cols-1 gap-x-8 gap-y-4 sm:grid-cols-2">
-                      <form.Field name="runs_on">
-                        {(field) => <RunnerField field={field} org={org} />}
-                      </form.Field>
-                    </div>
-                  )
+                  ) : null
                 }
               </form.Subscribe>
