+++ web/src/pages/assignments/formFieldHelpers.ts
@@ -30,3 +30,38 @@ export const normalizeOnBlur = (
     field.handleBlur()
   }
 }
+
+// Format a Date as a `datetime-local` input value (local wall-clock, no zone).
+export const toDatetimeLocalValue = (date: Date) => {
+  const pad = (value: number) => String(value).padStart(2, "0")
+
+  const year = date.getFullYear()
+  const month = pad(date.getMonth() + 1)
+  const day = pad(date.getDate())
+  const hours = pad(date.getHours())
+  const minutes = pad(date.getMinutes())
+
+  return `${year}-${month}-${day}T${hours}:${minutes}`
+}
+
+// Create-mode default: a week out gives students a sensible runway and avoids
+// the form defaulting to an already-overdue "now".
+export const sevenDaysFromNow = () => {
+  const date = new Date()
+  date.setDate(date.getDate() + 7)
+  return date
+}
+
+// Parse a stored UTC ISO instant into a `datetime-local` value; "" when absent
+// or unparseable.
+export const utcIsoToDatetimeLocalValue = (value?: string) => {
+  if (!value) return ""
+
+  const date = new Date(value)
+
+  if (Number.isNaN(date.getTime())) {
+    return ""
+  }
+
+  return toDatetimeLocalValue(date)
+}
